removing llamaindex
