+++ packages/cli/src/nodehost.ts
@@ -8,7 +8,6 @@ import { glob } from "glob"
 import { debug, error, info, warn } from "./log"
 import { execa } from "execa"
 import { join } from "node:path"
-import { LlamaIndexRetrievalService } from "./llamaindexretrieval"
 import { createNodePath } from "./nodepath"
 import { DockerManager } from "./docker"
 import { DefaultAzureCredential, AccessToken } from "@azure/identity"
@@ -24,11 +23,12 @@ import {
     AZURE_OPENAI_TOKEN_SCOPES,
     SHELL_EXEC_TIMEOUT,
     DOT_ENV_FILENAME,
+    MODEL_PROVIDER_OLLAMA,
+    TOOL_ID,
 } from "../../core/src/constants"
 import { createFileSystem, filterGitIgnore } from "../../core/src/fs"
 import {
     ServerManager,
-    RetrievalService,
     ModelService,
     LanguageModelConfiguration,
     LogLevel,
@@ -36,8 +36,12 @@ import {
     UTF8Encoder,
     RuntimeHost,
     setRuntimeHost,
+    ResponseStatus,
 } from "../../core/src/host"
-import { resolveLanguageModel } from "../../core/src/models"
+import {
+    parseModelIdentifier,
+    resolveLanguageModel,
+} from "../../core/src/models"
 import { createBundledParsers } from "../../core/src/pdf"
 import { AbortSignalOptions, TraceOptions } from "../../core/src/trace"
 import { unique } from "../../core/src/util"
@@ -51,10 +55,41 @@ class NodeServerManager implements ServerManager {
     }
 }
 
+class ModelManager implements ModelService {
+    constructor(private readonly host: RuntimeHost) {}
+    private async getModelToken(modelId: string) {
+        const { provider } = parseModelIdentifier(modelId)
+        const conn = await this.host.getLanguageModelConfiguration(modelId)
+        if (provider === MODEL_PROVIDER_OLLAMA)
+            conn.base = conn.base.replace(/\/v1$/i, "")
+        return conn
+    }
+
+    async pullModel(modelid: string): Promise<ResponseStatus> {
+        const { provider, model } = parseModelIdentifier(modelid)
+        const conn = await this.getModelToken(modelid)
+        if (provider === MODEL_PROVIDER_OLLAMA) {
+            const res = await fetch(`${conn.base}/api/pull`, {
+                method: "POST",
+                headers: {
+                    "user-agent": TOOL_ID,
+                    "content-type": "application/json",
+                },
+                body: JSON.stringify({ name: model, stream: false }, null, 2),
+            })
+            if (res.ok) {
+                const resp = await res.json()
+            }
+            return { ok: res.ok, status: res.status }
+        }
+
+        return { ok: true }
+    }
+}
+
 export class NodeHost implements RuntimeHost {
     readonly dotEnvPath: string
     userState: any = {}
-    retrieval: RetrievalService
     models: ModelService
     readonly path = createNodePath()
     readonly server = new NodeServerManager()
@@ -76,9 +111,7 @@ export class NodeHost implements RuntimeHost {
             })
             if (res.error) throw res.error
         }
-        const srv = new LlamaIndexRetrievalService(this)
-        this.retrieval = srv
-        this.models = srv
+        this.models = new ModelManager(this)
     }
 
     static async install(dotEnvPath: string) {
