+++ packages/cli/src/retrieval.ts
@@ -1,57 +1,8 @@
 import { expandFiles } from "../../core/src/fs"
 import { fuzzSearch } from "../../core/src/fuzzsearch"
-import {
-    upsertVector,
-    clearVectorIndex,
-    vectorSearch,
-} from "../../core/src/retrieval"
 import { normalizeFloat, normalizeInt } from "../../core/src/util"
 import { YAMLStringify } from "../../core/src/yaml"
 import { createProgressSpinner } from "./spinner"
-
-export async function retrievalIndex(
-    files: string[],
-    options: {
-        excludedFiles: string[]
-        name: string
-        model: string
-        temperature: string
-        chunkSize: string
-        chunkOverlap: string
-    }
-) {
-    const {
-        excludedFiles,
-        name: indexName,
-        model,
-        temperature,
-        chunkOverlap,
-        chunkSize,
-    } = options || {}
-    const fs = await expandFiles(files, excludedFiles)
-    if (!fs.length) {
-        console.error("no files matching")
-        return
-    }
-
-    const progress = createProgressSpinner(`indexing ${fs.length} files`)
-    await upsertVector(fs, {
-        progress,
-        indexName,
-        embedModel: model,
-        temperature: normalizeFloat(temperature),
-        chunkOverlap: normalizeInt(chunkOverlap),
-        chunkSize: normalizeInt(chunkSize),
-    })
-}
-
-export async function retrievalClear(options: {
-    name: string
-    summary: boolean
-}) {
-    const { name: indexName, summary } = options || {}
-    await clearVectorIndex({ indexName })
-}
 
 export async function retrievalSearch(
     q: string,
