+++ packages/core/src/constants.ts
@@ -44,12 +44,6 @@ export const TRACE_NODE_PREFIX = "genaiscript/trace/"
 export const EXTENSION_ID = "genaiscript.genaiscript-vscode"
 export const CHAT_PARTICIPANT_ID = TOOL_ID
 export const BING_SEARCH_ENDPOINT = "https://api.bing.microsoft.com/v7.0/search"
-export const LLAMAINDEX_SIMILARITY_TOPK = 5
-export const LLAMAINDEX_MIN_SCORE = 0.1
-export const RETRIEVAL_DEFAULT_INDEX = "default"
-export const RETRIEVAL_DEFAULT_LLM_MODEL = "gpt-35-turbo"
-export const RETRIEVAL_DEFAULT_EMBED_MODEL = "text-embedding-ada-002"
-export const RETRIEVAL_DEFAULT_TEMPERATURE = 0
 export const SYSTEM_FENCE = "\n"
 export const MAX_DATA_REPAIRS = 1
 export const NPM_CLI_PACKAGE = "genaiscript"
