+++ packages/core/src/host.ts
@@ -50,14 +50,8 @@ export interface ResponseStatus {
 }
 
 export interface RetrievalSearchOptions extends VectorSearchOptions {
-    files?: string[]
     topK?: number
     minScore?: number
-}
-
-export interface RetrievalUpsertOptions extends VectorSearchEmbeddingsOptions {
-    content?: string
-    mimeType?: string
 }
 
 export interface RetrievalSearchResponse extends ResponseStatus {
@@ -69,14 +63,9 @@ export interface ModelService {
 }
 
 export interface RetrievalService {
-    init(trace?: MarkdownTrace): Promise<void>
-    vectorClear(options?: VectorSearchOptions): Promise<ResponseStatus>
-    vectorUpsert(
-        filenameOrUrl: string,
-        options?: RetrievalUpsertOptions
-    ): Promise<ResponseStatus>
     vectorSearch(
         text: string,
+        files: WorkspaceFile[],
         options?: RetrievalSearchOptions
     ): Promise<RetrievalSearchResponse>
 }
@@ -155,7 +144,6 @@ export interface Host {
 }
 
 export interface RuntimeHost extends Host {
-    retrieval: RetrievalService
     models: ModelService
     workspace: Omit<WorkspaceFileSystem, "grep">
