+++ packages/core/src/server/messages.ts
@@ -1,11 +1,5 @@
 import { GenerationResult } from "../expander"
-import {
-    ParsePdfResponse,
-    ResponseStatus,
-    RetrievalSearchOptions,
-    RetrievalSearchResponse,
-    RetrievalUpsertOptions as RetrievalVectorUpsertOptions,
-} from "../host"
+import { ParsePdfResponse, ResponseStatus } from "../host"
 
 export interface RequestMessage {
     type: string
